+++ README.md
@@ -40,12 +40,6 @@
 - [x] CORS 설정하기
 - [x] CORS 설정 테스트 
 
-## https 통신을 위한 설정
-- [ ] DuckDns를 이용하여 도메인 설정하기
-- [ ] Nginx를 이용하여 https 설정하기
-- [ ] Cerbot을 이용하여 SSL 인증서 발급하기
-- [ ] OpenSSL을 이용하여 SSL 인증서 발급하기
-
 --- 
 # 6주차 3단계 구현해야할 기능 목록들 정리
 ## 1. 회원가입 API 수정하기
@@ -62,11 +56,3 @@
 
 
 ---
-# 6주차 4단계 구현해야할 기능 목록들 정리
-### 추가적으로 생각해봐야 할 API들
-- 카카오 로그인 API 
-- 특정 카테고리 조회, 새로운 카테고리 추가, 카테고리 삭제 API
-- 옵션 추가 및 삭제 API
-- 주문 추가 API
-- 상품 추가, 상품 정보 수정, 상품 삭제 API
-
