+++ src/main/java/gift/member/service/MemberService.java
@@ -6,8 +6,6 @@ public interface MemberService {
 
     void registerMember(String email, String password);
 
-    String generateToken(Member member);
-
     String login(String email, String password);
 
     Long getPoint(String memberEmail);
